move DrawerCloseButton to constants.ts
