--- packages/vlossom/src/components/vs-drawer/stories/DrawerCloseBotton.ts
@@ -1,8 +1,0 @@
-import { VsIcon } from '@/icons';
-
-export const DrawerCloseButton = {
-    components: { VsIcon },
-    template: `
-        <vs-icon icon="close" style="cursor: pointer" />
-    `,
-};
