+++ packages/vlossom/src/components/vs-drawer/stories/constants.ts
@@ -1,3 +1,12 @@
+import { VsIcon } from '@/icons';
+
+export const DrawerCloseButton = {
+    components: { VsIcon },
+    template: `
+        <vs-icon icon="close" style="cursor: pointer" />
+    `,
+};
+
 const contentStyle = {
     display: 'flex',
     justifyContent: 'space-between',
